Move more code over from retroarch.c to input_driver.c - move more internal state over
